FriendsQuestions listing working The only other action is MakePublic, it not working because of the way the id parameter is fetched.

+++ trunk/QuestionYourFriends/QuestionYourFriends/BusinessManagement/Question.cs
@@ -98,7 +98,7 @@ namespace QuestionYourFriends.BusinessManagement
             return QuestionYourFriendsDataAccess.DataAccess.Question.GetListOfReceiver(Context.QyfEntities, id);
         }
         
-        public static  List<QuestionYourFriendsDataAccess.Question> GetFriendsQuestions(long[] friends)
+        public static  List<QuestionYourFriendsDataAccess.Question> GetFriendsQuestions(int[] friends)
         {
             return QuestionYourFriendsDataAccess.DataAccess.Question.GetFriendsQuestions(Context.QyfEntities, friends);
         }
